Green the ruff CI lint + fix WatchlistItem alert cascade
- Clear all 129 pre-existing `ruff check app/` findings so the CI lint step
  passes: wrapped 53 over-length lines by hand (no noqa), removed unused
  imports/vars, snake_cased locals, isinstance-union, contextlib.suppress,
  ternaries, and other safe rewrites — all behaviour-preserving.
- WatchlistItem.alerts: dropped delete-orphan (same bug as Holding.alerts) so
  deleting a watchlist item detaches its alerts (DB SET NULL) instead of
  deleting the user's alerts. Added a regression test.

Verified: `ruff check app/` exits 0, 405 backend tests pass (1 new), no
config/rule was relaxed to achieve it.

+++ backend/app/services/recurring_detection_service.py
@@ -148,7 +148,7 @@ async def detect_recurring(
 
         holding = holdings[holding_id]
         last_date = dates[-1]
-        next_expected = last_date + timedelta(days=int(round(avg_interval)))
+        next_expected = last_date + timedelta(days=round(avg_interval))
 
         detected.append(
             {
